Allow limiting the number of outstanding prefetch requests
Currently prefetch will send all requests for all chunks for the file in
one shot. Whith the default chunk size of 32k, this can result in many
thousand outstanding requests for large files. Some servers like Serv-U
15.2.3.742 seem to be dropping requests after a certain number, which
results in the file download hanging indefinitely (or until the server
closes the connection).

Fix this issue by letting the user specify a limit for the number of
outstanding requests. This is similar to openssh's sftp, which limits
the number of outstanding requests to 64 by default.