Make test for prefetch requests limit stable
The prefetch requests are added in a background process, so the might
not be ready when the test wants to count them. Fix it by introducing a
wait_for untility function that waits for a assert to pass for a
specified timeout, to give the background thread the chance to start and
add the prefetch requests.

Also lock the prefetch lock before trying to access the extents to
prevent a data race.